Use storybook env config for creating package version env variable
