+++ packages/docs/vite.config.ts
@@ -5,7 +5,6 @@ import path from 'path';
 import { defineConfig } from 'vite';
 import { replaceCodePlugin } from 'vite-plugin-replace';
 import VitePluginCustomElementsManifest from 'vite-plugin-cem';
-import VitePluginPackageVersions from './src/vite-plugin-package-versions/index.js';
 import packageJson from '../components/package.json';
 import customElementConfig from '../components/custom-elements-manifest.config.js';
 import vitePluginSynergyStyles from './src/vite-plugin-synergy-styles/index.js';
@@ -93,9 +92,6 @@ export default defineConfig(() => ({
       files: ['../components/src/components/**/*.component.ts'],
       lit: true,
       plugins: getCustomElementManifestPlugins(),
-    }),
-    VitePluginPackageVersions({
-      packagePaths: ['../components', '../assets', '../styles', '../tokens'],
     }),
     replaceCodePlugin({
       replacements: [{
